+++ src/aimanager/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+1.5.3b1
+++++++
+* ``az aimanager list`` and ``az aimanager show``: Improve ``-o table`` output. Drop the
+  ``ETag`` column and add ``ProvisioningState`` and ``Subscription`` columns.
+
 1.5.2b1
 ++++++
 * Refactor validation code to make the name validators consistent